+++ CHANGELOG.md
@@ -16,6 +16,10 @@ All notable changes to this project will be documented in this file.
 - Throw a dedicated exception on empty parameters when authenticating to a docker registry. (#579)
 - Check replicate recoverability wih a dedicated check in `ReplicateRecoveryService`. (#580)
 
+### Dependency Upgrades
+
+- Upgrade to `iexec-common` 8.4.0. (#583)
+
 ## [[8.3.0]](https://github.com/iExecBlockchainComputing/iexec-worker/releases/tag/v8.3.0) 2024-01-11
 
 ### New Features